refactor(tests): split test_squad_tools.py per agent
The 1164-line test_squad_tools.py is a top-3 offender in the context-bloat
review on #114. The file is structured class-per-agent and every test
import is method-local, so the split is mechanical: each class moves into
tests/test_squad_<agent>.py with its module docstring scoped to that
agent. No test bodies, fixtures, or symbol names change.

Sizes after the split:
  test_squad_workspace_tools.py            43
  test_squad_programme_manager.py          60
  test_squad_disclosure_coordinator.py     62
  test_squad_penetration_tester.py        184
  test_squad_technical_author.py          202
  test_squad_vulnerability_researcher.py  319
  test_squad_osint_analyst.py             358

pytest -m unit --cov: 902 passed, 92.24% coverage.
