+++ tests/test_squad_technical_author.py
@@ -1,0 +1,202 @@
+"""
+tests/test_squad_technical_author.py - exercise the @tool wrappers on the
+Technical Author.
+
+The wrappers are thin: unmarshal JSON, call into tools/* helpers, serialise
+the result. Coverage here is regression coverage of the wrapping itself; the
+underlying helpers are exercised in their own dedicated test files.
+"""
+
+from __future__ import annotations
+
+import json
+from unittest.mock import patch
+
+import pytest
+
+pytestmark = pytest.mark.unit
+
+
+class TestTechnicalAuthorTools:
+    @staticmethod
+    def _write_verified(tmp_path, verified_vuln) -> None:
+        (tmp_path / "verified.json").write_text(
+            json.dumps([verified_vuln.model_dump(mode="json")]),
+            encoding="utf-8",
+        )
+
+    @staticmethod
+    def _good_authoring(**overrides):
+        base = {
+            "finding_index": 0,
+            "title": "SQL Injection in /search?q allows full database extraction",
+            "summary": (
+                "The /search endpoint concatenates user input into a SELECT statement. "
+                "An unauthenticated attacker can dump the entire users table."
+            ),
+            "description": (
+                "The handler at routes/search.py concatenates the q parameter directly "
+                "into the SQL statement with no parameterisation. Standard UNION-based "
+                "injection extracts arbitrary rows from the users table."
+            ),
+            "steps_to_reproduce": [
+                "Issue GET https://api.example.com/search?q=test' UNION SELECT 1,2,3-- ",
+                "Observe the response body contains the union'd rows.",
+            ],
+            "evidence": 'HTTP/1.1 200 OK\n\n[{"username":"alice"}]',
+            "impact": (
+                "An unauthenticated attacker can dump the entire users table including "
+                "bcrypt hashes and email addresses, enabling offline cracking and full "
+                "account takeover."
+            ),
+            "remediation": (
+                "Use parameterised queries throughout the ORM. See "
+                "https://cheatsheetseries.owasp.org/cheatsheets/SQL_Injection_Prevention_Cheat_Sheet.html"
+            ),
+            "cvss_vector": "CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H",
+            "cwe_id": 89,
+        }
+        base.update(overrides)
+        return base
+
+    def test_draft_report_tool_writes_draft_and_returns_validation(
+        self, verified_vuln, tmp_path
+    ) -> None:
+        from squad.technical_author import draft_report_tool
+
+        self._write_verified(tmp_path, verified_vuln)
+        with (
+            patch("tools.workspace.runtime.run_dir", return_value=tmp_path),
+            patch("tools.report_tools.runtime.run_dir", return_value=tmp_path),
+        ):
+            result = draft_report_tool.func(**self._good_authoring())
+
+        assert isinstance(result, dict)
+        assert result["validation"]["ok"] is True
+        assert (tmp_path / "drafts" / "000.json").exists()
+
+    def test_draft_report_tool_surfaces_validation_issues(self, verified_vuln, tmp_path) -> None:
+        from squad.technical_author import draft_report_tool
+
+        self._write_verified(tmp_path, verified_vuln)
+        with (
+            patch("tools.workspace.runtime.run_dir", return_value=tmp_path),
+            patch("tools.report_tools.runtime.run_dir", return_value=tmp_path),
+        ):
+            result = draft_report_tool.func(**self._good_authoring(title="bad title"))
+
+        assert isinstance(result, dict)
+        assert result["validation"]["ok"] is False
+        sections = {i["section"] for i in result["validation"]["issues"]}
+        assert "title" in sections
+
+    def test_draft_report_tool_rejects_out_of_range_index(self, verified_vuln, tmp_path) -> None:
+        from squad.technical_author import draft_report_tool
+
+        self._write_verified(tmp_path, verified_vuln)
+        with (
+            patch("tools.workspace.runtime.run_dir", return_value=tmp_path),
+            patch("tools.report_tools.runtime.run_dir", return_value=tmp_path),
+        ):
+            with pytest.raises(ValueError, match="out of range"):
+                draft_report_tool.func(**self._good_authoring(finding_index=5))
+
+    def test_finalise_reports_tool_consolidates_drafts(self, verified_vuln, tmp_path) -> None:
+        from squad.technical_author import draft_report_tool, finalise_reports_tool
+
+        self._write_verified(tmp_path, verified_vuln)
+        with (
+            patch("tools.workspace.runtime.run_dir", return_value=tmp_path),
+            patch("tools.report_tools.runtime.run_dir", return_value=tmp_path),
+        ):
+            draft_report_tool.func(**self._good_authoring())
+            result = finalise_reports_tool.func("acme", "Session summary line.")
+
+        assert result == "reports.json"
+        assert (tmp_path / "reports.json").exists()
+
+    def test_finalise_reports_tool_raises_on_unresolved_errors(
+        self, verified_vuln, tmp_path
+    ) -> None:
+        from squad.technical_author import draft_report_tool, finalise_reports_tool
+
+        self._write_verified(tmp_path, verified_vuln)
+        with (
+            patch("tools.workspace.runtime.run_dir", return_value=tmp_path),
+            patch("tools.report_tools.runtime.run_dir", return_value=tmp_path),
+        ):
+            draft_report_tool.func(**self._good_authoring(title="bad title"))
+            with pytest.raises(ValueError, match="unresolved errors"):
+                finalise_reports_tool.func("acme", "Summary.")
+
+    def test_sanitise_evidence_tool_returns_redactions(self) -> None:
+        from squad.technical_author import sanitise_evidence_tool
+
+        result = sanitise_evidence_tool.func("Authorization: Bearer abc.def.ghi")
+        assert isinstance(result, dict)
+        assert "Bearer abc.def.ghi" not in result["sanitised"]
+        assert result["redactions"]
+
+    def test_lookup_cwe_tool_finds_known_class(self) -> None:
+        from squad.technical_author import lookup_cwe_tool
+
+        result = lookup_cwe_tool.func("SQLi")
+        assert isinstance(result, list)
+        assert result
+        assert result[0]["cwe_id"] == 89
+        assert "cwe.mitre.org" in result[0]["url"]
+
+    def test_lookup_cwe_tool_empty_for_unknown(self) -> None:
+        from squad.technical_author import lookup_cwe_tool
+
+        assert lookup_cwe_tool.func("zzz-not-a-class") == []
+
+    def test_lookup_owasp_tool_returns_cheatsheet(self) -> None:
+        from squad.technical_author import lookup_owasp_tool
+
+        result = lookup_owasp_tool.func("sql injection")
+        assert isinstance(result, list)
+        assert result
+        assert any("SQL_Injection_Prevention" in r["url"] for r in result)
+
+    def test_calculate_cvss_tool(self) -> None:
+        from squad.technical_author import calculate_cvss_tool
+
+        with patch(
+            "squad.technical_author.calculate_cvss_score",
+            return_value=8.8,
+        ) as m:
+            result = calculate_cvss_tool.func("CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H")
+
+        assert result == 8.8
+        m.assert_called_once_with("CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H")
+
+    def test_list_programme_reports_tool(self) -> None:
+        from squad.technical_author import list_programme_reports_tool
+
+        h1_reports = [
+            {
+                "id": "1",
+                "attributes": {
+                    "title": "Existing report",
+                    "severity_rating": "high",
+                    "state": "triaged",
+                },
+            }
+        ]
+        with (
+            patch("squad.technical_author.http.set_programme") as mhttp,
+            patch("squad.technical_author.h1.list_reports", return_value=h1_reports) as mlist,
+        ):
+            result = list_programme_reports_tool.func("acme", page_size=10)
+
+        assert result == [
+            {
+                "report_id": "1",
+                "title": "Existing report",
+                "severity": "high",
+                "state": "triaged",
+            }
+        ]
+        mhttp.assert_called_once_with("acme")
+        mlist.assert_called_once_with("acme", page_size=10)
