+++ tests/test_squad_vulnerability_researcher.py
@@ -1,0 +1,319 @@
+"""
+tests/test_squad_vulnerability_researcher.py - exercise the @tool wrappers on
+the Vulnerability Researcher.
+
+The wrappers are thin: unmarshal JSON, call into tools/* helpers, serialise
+the result. Coverage here is regression coverage of the wrapping itself; the
+underlying helpers are exercised in their own dedicated test files.
+"""
+
+from __future__ import annotations
+
+import json
+from unittest.mock import patch
+
+import pytest
+
+pytestmark = pytest.mark.unit
+
+
+class TestVulnerabilityResearcherTools:
+    def test_lookup_cve_tool(self) -> None:
+        from squad.vulnerability_researcher import lookup_cve_tool
+
+        sentinel = [{"cve_id": "CVE-2024-1234"}]
+        with patch(
+            "squad.vulnerability_researcher.lookup_cve",
+            return_value=sentinel,
+        ) as m:
+            result = lookup_cve_tool.func("sql injection")
+
+        assert result == sentinel
+        m.assert_called_once_with("sql injection")
+
+    def test_list_programme_reports_tool_returns_all(self) -> None:
+        from squad.vulnerability_researcher import list_programme_reports_tool
+
+        reports = [
+            {
+                "id": "1",
+                "attributes": {
+                    "title": "SQL Injection in search",
+                    "severity_rating": "high",
+                    "state": "open",
+                },
+            },
+            {
+                "id": "2",
+                "attributes": {
+                    "title": "XSS in admin panel",
+                    "severity_rating": "medium",
+                    "state": "triaged",
+                },
+            },
+        ]
+        with (
+            patch("squad.vulnerability_researcher.http.set_programme"),
+            patch(
+                "squad.vulnerability_researcher.h1.list_reports",
+                return_value=reports,
+            ),
+        ):
+            result = list_programme_reports_tool.func("acme")
+
+        assert isinstance(result, list)
+        assert len(result) == 2
+        assert result[0]["report_id"] == "1"
+        assert result[1]["severity"] == "medium"
+
+    def test_list_programme_reports_tool_handles_missing_fields(self) -> None:
+        from squad.vulnerability_researcher import list_programme_reports_tool
+
+        reports = [{"id": "3", "attributes": {"state": "closed"}}]
+        with (
+            patch("squad.vulnerability_researcher.http.set_programme"),
+            patch(
+                "squad.vulnerability_researcher.h1.list_reports",
+                return_value=reports,
+            ),
+        ):
+            result = list_programme_reports_tool.func("acme")
+
+        assert isinstance(result, list)
+        assert len(result) == 1
+        assert result[0]["title"] is None
+        assert result[0]["severity"] is None
+
+    # Triage-task tools
+
+    @staticmethod
+    def _write_findings(tmp_path, raw_finding_high) -> None:
+        (tmp_path / "findings.json").write_text(
+            json.dumps([raw_finding_high.model_dump(mode="json")]),
+            encoding="utf-8",
+        )
+
+    @staticmethod
+    def _good_authoring(**overrides):
+        base = {
+            "finding_index": 0,
+            "severity_decision": "keep",
+            "severity": "high",
+            "severity_rationale": (
+                "Unauthenticated SQLi at a public endpoint with full DB read "
+                "available - matches the PT high call."
+            ),
+            "cvss_vector": "CVSS:3.1/AV:N/AC:L/PR:L/UI:N/S:U/C:H/I:H/A:H",
+            "title": "SQL Injection in /search?q allows database extraction",
+            "description": (
+                "The /search endpoint concatenates q into a SELECT statement without "
+                "parameterisation. sqlmap exploits classic UNION-based injection to "
+                "extract rows from the users table."
+            ),
+            "steps_to_reproduce": [
+                "GET https://api.example.com/search?q=test' UNION SELECT 1,2,3-- ",
+                "Observe the response body contains the union'd rows.",
+            ],
+            "impact": (
+                "An authenticated attacker dumps the users table including bcrypt "
+                "hashes and emails, enabling offline cracking and account takeover."
+            ),
+            "remediation": (
+                "Use parameterised queries throughout. See "
+                "https://cheatsheetseries.owasp.org/cheatsheets/SQL_Injection_Prevention_Cheat_Sheet.html"
+            ),
+            "programme_handle": "test-programme",
+        }
+        base.update(overrides)
+        return base
+
+    def test_list_raw_findings_tool(self, raw_finding_high, tmp_path) -> None:
+        from squad.vulnerability_researcher import list_raw_findings_tool
+
+        self._write_findings(tmp_path, raw_finding_high)
+        with patch("tools.workspace.runtime.run_dir", return_value=tmp_path):
+            result = list_raw_findings_tool.func()
+
+        assert isinstance(result, list)
+        assert result[0]["index"] == 0
+        assert result[0]["vuln_class"] == raw_finding_high.vuln_class
+        assert result[0]["evidence_bytes"] == len(raw_finding_high.evidence)
+
+    def test_read_raw_finding_tool(self, raw_finding_high, tmp_path) -> None:
+        from squad.vulnerability_researcher import read_raw_finding_tool
+
+        self._write_findings(tmp_path, raw_finding_high)
+        with patch("tools.workspace.runtime.run_dir", return_value=tmp_path):
+            result = read_raw_finding_tool.func(0)
+
+        assert isinstance(result, dict)
+        assert result["target"] == raw_finding_high.target
+
+    def test_read_raw_finding_tool_rejects_out_of_range(self, raw_finding_high, tmp_path) -> None:
+        from squad.vulnerability_researcher import read_raw_finding_tool
+
+        self._write_findings(tmp_path, raw_finding_high)
+        with patch("tools.workspace.runtime.run_dir", return_value=tmp_path):
+            with pytest.raises(ValueError, match="out of range"):
+                read_raw_finding_tool.func(5)
+
+    def test_lookup_cwe_tool_finds_known_class(self) -> None:
+        from squad.vulnerability_researcher import lookup_cwe_tool
+
+        result = lookup_cwe_tool.func("SQLi")
+        assert isinstance(result, list)
+        assert result
+        assert result[0]["cwe_id"] == 89
+
+    def test_lookup_owasp_tool_returns_cheatsheet(self) -> None:
+        from squad.vulnerability_researcher import lookup_owasp_tool
+
+        result = lookup_owasp_tool.func("sql injection")
+        assert isinstance(result, list)
+        assert any("SQL_Injection_Prevention" in r["url"] for r in result)
+
+    def test_calculate_cvss_tool(self) -> None:
+        from squad.vulnerability_researcher import calculate_cvss_tool
+
+        result = calculate_cvss_tool.func("CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H")
+        assert result == 9.8
+
+    @staticmethod
+    def _patch_programme(programme):
+        return [
+            patch("squad.vulnerability_researcher.http.set_programme"),
+            patch(
+                "squad.vulnerability_researcher.h1.get_programme_policy",
+                return_value={"data": {}},
+            ),
+            patch("squad.vulnerability_researcher.h1.get_structured_scope", return_value={}),
+            patch(
+                "squad.vulnerability_researcher.h1.parse_programme",
+                return_value=programme,
+            ),
+        ]
+
+    def test_assess_finding_tool_writes_assessment_and_returns_validation(
+        self, raw_finding_high, programme, tmp_path
+    ) -> None:
+        from squad.vulnerability_researcher import assess_finding_tool
+
+        self._write_findings(tmp_path, raw_finding_high)
+        patches = self._patch_programme(programme) + [
+            patch("tools.workspace.runtime.run_dir", return_value=tmp_path),
+            patch("tools.triage_tools.runtime.run_dir", return_value=tmp_path),
+        ]
+        for p in patches:
+            p.start()
+        try:
+            result = assess_finding_tool.func(**self._good_authoring())
+        finally:
+            for p in reversed(patches):
+                p.stop()
+
+        assert isinstance(result, dict)
+        assert result["validation"]["ok"] is True
+        assert (tmp_path / "assessments" / "000.json").exists()
+
+    def test_assess_finding_tool_surfaces_validation_issues(
+        self, raw_finding_high, programme, tmp_path
+    ) -> None:
+        from squad.vulnerability_researcher import assess_finding_tool
+
+        self._write_findings(tmp_path, raw_finding_high)
+        patches = self._patch_programme(programme) + [
+            patch("tools.workspace.runtime.run_dir", return_value=tmp_path),
+            patch("tools.triage_tools.runtime.run_dir", return_value=tmp_path),
+        ]
+        for p in patches:
+            p.start()
+        try:
+            result = assess_finding_tool.func(**self._good_authoring(description="Too short."))
+        finally:
+            for p in reversed(patches):
+                p.stop()
+
+        assert isinstance(result, dict)
+        assert result["validation"]["ok"] is False
+        sections = {i["section"] for i in result["validation"]["issues"]}
+        assert "description" in sections
+
+    def test_discard_finding_tool_writes_discard(self, raw_finding_high, tmp_path) -> None:
+        from squad.vulnerability_researcher import discard_finding_tool
+
+        self._write_findings(tmp_path, raw_finding_high)
+        with (
+            patch("tools.workspace.runtime.run_dir", return_value=tmp_path),
+            patch("tools.triage_tools.runtime.run_dir", return_value=tmp_path),
+        ):
+            result = discard_finding_tool.func(
+                finding_index=0,
+                reason="false_positive",
+                rationale="On reproduction the tool's marker did not appear.",
+            )
+
+        assert isinstance(result, dict)
+        assert (tmp_path / "discards" / "000.json").exists()
+
+    def test_discard_finding_tool_rejects_thin_rationale(self, raw_finding_high, tmp_path) -> None:
+        from squad.vulnerability_researcher import discard_finding_tool
+
+        self._write_findings(tmp_path, raw_finding_high)
+        with (
+            patch("tools.workspace.runtime.run_dir", return_value=tmp_path),
+            patch("tools.triage_tools.runtime.run_dir", return_value=tmp_path),
+        ):
+            with pytest.raises(ValueError, match="at least 10 chars"):
+                discard_finding_tool.func(
+                    finding_index=0, reason="false_positive", rationale="nope"
+                )
+
+    def test_finalise_triage_tool_consolidates_assessments(
+        self, raw_finding_high, programme, tmp_path
+    ) -> None:
+        from squad.vulnerability_researcher import assess_finding_tool, finalise_triage_tool
+
+        self._write_findings(tmp_path, raw_finding_high)
+        patches = self._patch_programme(programme) + [
+            patch("tools.workspace.runtime.run_dir", return_value=tmp_path),
+            patch("tools.triage_tools.runtime.run_dir", return_value=tmp_path),
+        ]
+        for p in patches:
+            p.start()
+        try:
+            assess_finding_tool.func(**self._good_authoring())
+            result = finalise_triage_tool.func("test-programme")
+        finally:
+            for p in reversed(patches):
+                p.stop()
+
+        assert result == "verified.json"
+        assert (tmp_path / "verified.json").exists()
+
+    def test_finalise_triage_tool_raises_on_unprocessed(
+        self, raw_finding_high, programme, tmp_path
+    ) -> None:
+        from squad.vulnerability_researcher import finalise_triage_tool
+
+        # write TWO raw findings but assess none
+        (tmp_path / "findings.json").write_text(
+            json.dumps(
+                [
+                    raw_finding_high.model_dump(mode="json"),
+                    raw_finding_high.model_dump(mode="json"),
+                ]
+            ),
+            encoding="utf-8",
+        )
+        patches = self._patch_programme(programme) + [
+            patch("tools.workspace.runtime.run_dir", return_value=tmp_path),
+            patch("tools.triage_tools.runtime.run_dir", return_value=tmp_path),
+        ]
+        for p in patches:
+            p.start()
+        try:
+            with pytest.raises(ValueError, match="no assessment or discard"):
+                finalise_triage_tool.func("test-programme")
+        finally:
+            for p in reversed(patches):
+                p.stop()
